[Back-front] modif add-delete typesalle + salle + animation

+++ Sources/gestion-culturelle-back/src/main/java/com/afpa/cda/service/impl/AnimationServiceImpl.java
@@ -35,12 +35,12 @@ public class AnimationServiceImpl implements IAnimationService {
 	public AnimationDto findById(int id) {
 
 		Optional<Animation> animOp = this.animationRepository.findById(id);
-		AnimationDto animDto = null;
+		AnimationDto animationDto = null;
 		if (animOp.isPresent()) {
 			Animation an = animOp.get();
-			animDto = this.modelMapper.map(an,AnimationDto.class);
+			animationDto = this.modelMapper.map(an,AnimationDto.class);
 		}
-		return animDto;
+		return animationDto;
 	}
 
 
@@ -56,27 +56,35 @@ public class AnimationServiceImpl implements IAnimationService {
 
 
 	@Override
-	public AnimationDto add(AnimationDto anim) {
-		Animation animE = this.modelMapper.map(anim,Animation.class);
-		Animation animationEntity = this.animationRepository.save(animE);
-		anim.setId(animationEntity .getId());
-		return anim;
+	public boolean add(AnimationDto animationDto) {
+		Optional <Animation> animationOp = this.animationRepository.findByLabel(animationDto.getLabel());
+
+		if (!animationOp.isPresent()) {
+			this.animationRepository.save(this.modelMapper.map(animationDto,Animation.class));
+			return false;
+		}
+
+//		Animation animE = this.modelMapper.map(animationDto,Animation.class);
+//		Animation animationEntity = this.animationRepository.save(animE);
+		//animationDto.setId(animationEntity .getId());
+//		return false;
+		return true;
 	}
 
 
 	@Override
-	public boolean update(AnimationDto animDto, int id) {
-		Optional<Animation> animEU = this.animationRepository.findById(id);
-		if (animEU.isPresent()) {
-			Animation an = animEU.get();
-			an.setLabel(animDto.getLabel());
-			an.setType(animDto.getType());
-			an.setPrix(animDto.getPrix());
-			an.setNbreSpectateursPrevus(animDto.getNbreSpectateursPrevus());
+	public boolean update(AnimationDto animationDto, int id) {
+		Optional<Animation> animationOp = this.animationRepository.findById(id);
+		if (animationOp.isPresent()) {
+			Animation animation = animationOp.get();
+			animation.setLabel(animationDto.getLabel());
+			animation.setType(animationDto.getType());
+			animation.setPrix(animationDto.getPrix());
+			animation.setNbreSpectateursPrevus(animationDto.getNbreSpectateursPrevus());
 
 			List<Manifestation> listManifestation = manifestationRepository.findAll();
 			for (Manifestation manifestation : listManifestation) {
-				if (manifestation.getAnimation().getId()==an.getId()) {
+				if (manifestation.getAnimation().getId()==animation.getId()) {
 					ManifestationDto manifestationDto = modelMapper.map(manifestation,ManifestationDto.class);
 					manifestationDto=calcul(manifestationDto);
 					Manifestation manif = modelMapper.map(manifestationDto,Manifestation.class);
@@ -89,47 +97,53 @@ public class AnimationServiceImpl implements IAnimationService {
 
 			}
 
-			this.animationRepository.save(an);
+			this.animationRepository.save(animation);
 			return true;
 		}
 		return false;
 	}
 
 	@Override
-	public ManifestationDto calcul (ManifestationDto manifDto) {
+	public ManifestationDto calcul (ManifestationDto manifestationDto) {
 
-		double debut=manifDto.getDateDebut().getTime();
-		double fin=manifDto.getDateFin().getTime();
+		double debut=manifestationDto.getDateDebut().getTime();
+		double fin=manifestationDto.getDateFin().getTime();
 		double duree = 1+((((fin-debut)/1000)/3600)/24);
-		AnimationDto animDto = new AnimationDto();
-		Optional<Animation> animOp=animationRepository.findById(manifDto.getAnimation().getId());
-		if (animOp.isPresent()) {
-			Animation anim = animOp.get();
-			animDto=modelMapper.map(anim,AnimationDto.class);
+		AnimationDto animationDto = new AnimationDto();
+		Optional<Animation> animationOp=animationRepository.findById(manifestationDto.getAnimation().getId());
+		if (animationOp.isPresent()) {
+			Animation anim = animationOp.get();
+			animationDto=modelMapper.map(anim,AnimationDto.class);
 		}
 
 		SalleDto salleDto = new SalleDto ();
-		Optional<Salle> salleOp=salleRepository.findById(manifDto.getSalle().getId());
+		Optional<Salle> salleOp=salleRepository.findById(manifestationDto.getSalle().getId());
 		if (salleOp.isPresent()) {
 			Salle salle = salleOp.get();
 			salleDto = modelMapper.map(salle,SalleDto.class);
 		}
 
-		manifDto.setReservations(animDto.getNbreSpectateursPrevus());
-		manifDto.setReservationsVip(salleDto.getPlacesVip());
-		manifDto.setCout( (animDto.getPrix()+(duree* salleDto.getFraisJournalier())));
-		manifDto.setPrixBillet(manifDto.getCout()/(animDto.getNbreSpectateursPrevus()*0.8));
+		manifestationDto.setReservations(animationDto.getNbreSpectateursPrevus());
+		manifestationDto.setReservationsVip(salleDto.getPlacesVip());
+		manifestationDto.setCout( (animationDto.getPrix()+(duree* salleDto.getFraisJournalier())));
+		manifestationDto.setPrixBillet(manifestationDto.getCout()/(animationDto.getNbreSpectateursPrevus()*0.8));
 
-		return manifDto;
+		return manifestationDto;
 	}
 
 
 	@Override
 	public boolean delete(int id) {
-		if(this.animationRepository.existsById(id)) {
+		List <Manifestation> listManifestations = manifestationRepository.findManifestationByAnimationId(id);
+		
+		if (listManifestations.isEmpty() && this.animationRepository.existsById(id))		 {
 			this.animationRepository.deleteById(id);
 			return true;
 		}
+//		if(this.animationRepository.existsById(id)) {
+//			this.animationRepository.deleteById(id);
+//			return true;
+//		}
 		return false;
 	}
 }
